backend/code/requirements.txt
coverage
mock
fakeredis
pylint
flask
redis
google-cloud-profiler
opencensus.ext.stackdriver
grpcio
grpcio-tools